backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
google-genai
load_dotenv
